Add pretty printer for DecorationContainer.
It is based on the DecorablePrinter type in the mongodb/mongo repository
with some changes to leverage the libstdc++ GDB pretty printers more
directly.

The gdbmongo package is still lacking a convenient way to register its
pretty printer types (and hopefully avoid conflicting with ones from the
mongodb/mongo repository) but that will come later.