Add pretty printers for LockManager and related types.
Effectively implements mongo::LockManager::dump() by instead walking
in-memory data structures. This means the logic is compatible with core
dumps where it isn't possible to execute code (and has been tested
locally to confirm this).

Only supports relatively recents versions of the MongoDB Server
(probably 5.1 and later).

Lastly, PyLint is a little sensitive to the "duplication" of
`self.val = val` and display_hint() definitions in each of the pretty
printers. Bumping the minimum number of lines for similarity detection
from 4 to 6 seems to have quieted it down.