Fix bug with attempting to load our fake typing.Protocols in real GDB.
Changes PrettyPrinterProtocol, etc. to exist in GDB by having a concrete
version of the classes defined in the gdbmongo package as well. When
typing.TYPE_CHECKING is enabled, the version of the classes defined in
the stub files for the gdb module is used. This approach keeps the stub
files for the gdb module from depending on the gdbmongo package while
still keeping the code duplication in the gdbmongo package low.