Update a few type annotations.
Arguments to functions and methods in the gdb module have mostly been
declared as positional-only arguments up to this point. The idea behind
doing so is to lessen the likelihood of a compatibility issue in the
gdbmongo package if a newer version of GDB chooses to use a different
name as a keyword argument.