Optimize memory usage when pretty printing nested objects and arrays.
Let's say an object or array has objsize=NumBytes. Pretty printing all
of the children of such an object or array would otherwise consume
O(NumBytes * NumBytes) memory if the object or array contains deeply
nested objects or arrays.

This optimization won't apply to the mongo::BSONObj contained within the
synthetic MongoBSONCodeWScope values but BSON type 0x0F is deprecated
and is unlikely to be used in testing anyway.