Stop overwriting the executable file when using objcopy.
The changes from SERVER 48203 changed the default options for building
the MongoDB Server to --install-action=hardlink. The hard link cannot be
overwritten while the executable is running. When gdbmongo runs objcopy
to detect the compiler version from the running executable, it fails
with a "Text file busy" error and causes the libstdc++ pretty printers
to not be loaded.