Fix test_can_load_module_from_toolchain() test case not being run.
Moves the test_can_load_module_from_toolchain() function outside the
TestStdlibPrinters class.

The TestStdlibPrinters.test_can_load_module_from_toolchain() function
was being collected as a test case but was not being parameterized and
called. In pytest 7.2.0, a PytestReturnNotNoneWarning warning started to
be emitted due to the test function returning a MarkDecorator. It
appears pytest.mark.parametrize() and staticmethod() do not compose with
each other.