Fix AssertionError from gdb_resolve_type() when printing Abseil map.
The attributes gdb.Type.tag and gdb.Type.name are both None for function
types and so there is no other type to look up.

This problem got noticed after the DocumentSourceVisitorRegistry
decoration on ServiceContext was recently introduced by SERVER 62027 and
from it storing function pointers as its map values.