Fix cast of `unsigned char` value to decoration_type.
For decorations of a pointer type, such a mongo::VectorClock*, the
previous behavior caused a zeroish value to be displayed in GDB for the
decoration. This is because it isn't valid to cast a non-pointer type to
a pointer type. We instead cast an `unsigned char*` to a `VectorClock**`
to always be casting from a pointer type.