Disable formatting on _version.py file generated by setuptools_scm.
The 'setuptools_scm >= 8.0.0' package now includes a typing.Tuple type
annotation for the `__version_tuple__` variable which may cause the
100-character line length limit to be exceeded and for YAPF to generate
a diff for the file.

This exclusion setting is **not** maintained through the pyproject.toml
file with logic similar to the following:

    [tool.yapfignore]
    ignore_patterns = [
        "gdbmongo/_version.py"
    ]

because the lack of auto-enumeration in YAPF for .pyi files led
test_formatting.py to maintain its own globs of files to include.