Display LockerImpl's _threadId and _debugInfo in LockManager dump.
The thread's name and number (name assigned by the process itself and
number assigned by GDB) are also included in the LockManager dump when
the associated thread has been identified.

This information is pertinent when determining which operations are
involved in a server hang. Showing it directly in the LockManager dump
is more friendly than requiring an engineer to cast+dereference the
Locker* address and manually match up the mongo::LockerImpl::_threadId
value with the output from the `info threads` command.